+++ juniper/CHANGELOG.md
@@ -16,6 +16,7 @@ All user visible changes to `juniper` crate will be documented in this file. Thi
     - Made `includeDeprecated` argument of `__Type.fields`, `__Type.enumValues`, `__Type.inputFields`, `__Field.args` and `__Directive.args` fields non-`Null`. ([#1348], [graphql/graphql-spec#1142])
     - Made `@deprecated(reason:)` argument non-`Null`. ([#1348], [graphql/graphql-spec#1040])
 - Changed `ScalarToken::String` to contain raw quoted and escaped `StringLiteral` (was unquoted but escaped string before). ([#1349])
+- Added `LexerError::UnterminatedBlockString` variant. ([#1349])
 
 ### Added
 
@@ -29,6 +30,7 @@ All user visible changes to `juniper` crate will be documented in this file. Thi
         - `includeDeprecated` argument to `__Type.inputFields`, `__Field.args` and `__Directive.args` fields.
         - `__InputValue.isDeprecated` and `__InputValue.deprecationReason` fields.
         - `schema::meta::Argument::deprecation_status` field.
+- Support for [block strings][0180-1]. ([#1349])
 
 ### Changed
 
@@ -50,6 +52,7 @@ All user visible changes to `juniper` crate will be documented in this file. Thi
 [graphql/graphql-spec#825]: https://github.com/graphql/graphql-spec/pull/825
 [graphql/graphql-spec#1040]: https://github.com/graphql/graphql-spec/pull/1040
 [graphql/graphql-spec#1142]: https://github.com/graphql/graphql-spec/pull/1142
+[0180-1]: https://spec.graphql.org/September2025/#sec-String-Value.Block-Strings